requirements.txt
requests~=2.32.3
beautifulsoup4~=4.13.3
selenium~=4.30.0
urllib3~=2.3.0
tldextract~=5.1.3
lxml==5.3.1
google-generativeai==0.8.4
openai~=1.68.2
keybert~=0.9.0
streamlit~=1.43.2